+++ mobile/android/android-components/components/concept/tabstray/src/main/java/mozilla/components/concept/tabstray/Tab.kt
@@ -5,6 +5,7 @@
 package mozilla.components.concept.tabstray
 
 import android.graphics.Bitmap
+import mozilla.components.concept.engine.media.Media
 
 /**
  * Data class representing a tab to be displayed in a [TabsTray].
@@ -14,11 +15,13 @@ import android.graphics.Bitmap
  * @property title Current title of the tab (or an empty [String]]).
  * @property icon Current icon of the tab (or null)
  * @property thumbnail Current thumbnail of the tab (or null)
+ * @property mediaState Current media state for the tab (or null)
  */
 data class Tab(
     val id: String,
     val url: String,
     val title: String = "",
     val icon: Bitmap? = null,
-    val thumbnail: Bitmap? = null
+    val thumbnail: Bitmap? = null,
+    val mediaState: Media.State? = null
 )