+++ mobile/android/android-components/components/feature/tabs/src/main/java/mozilla/components/feature/tabs/ext/TabSessionState.kt
@@ -4,13 +4,20 @@
 
 package mozilla.components.feature.tabs.ext
 
+import mozilla.components.browser.state.state.MediaState
 import mozilla.components.browser.state.state.TabSessionState
+import mozilla.components.concept.engine.media.Media
 import mozilla.components.concept.tabstray.Tab
 
-internal fun TabSessionState.toTab() = Tab(
+internal fun TabSessionState.toTab(mediaState: MediaState.State) = Tab(
     id,
     content.url,
     content.title,
     content.icon,
-    content.thumbnail
+    content.thumbnail,
+    when (mediaState) {
+        MediaState.State.PLAYING -> Media.State.PLAYING
+        MediaState.State.PAUSED -> Media.State.PAUSED
+        else -> null
+    }
 )